+++ core/runner/src/RunnerSteps.ts
@@ -1,15 +1,24 @@
 import { type Context, GlobalError, type NodeBase, type Step } from "@nanoservice-ts/shared";
-import type NanoService from "./NanoService";
 import type NanoServiceResponse from "./NanoServiceResponse";
 
 export default abstract class RunnerSteps {
-	async runSteps(ctx: Context, steps: NanoService[], deep = false, step_name = ""): Promise<Context> {
+	/**
+	 * Executes a series of steps in the given context.
+	 *
+	 * @param ctx - The context in which the steps are executed.
+	 * @param steps - An array of NanoService steps to be executed.
+	 * @param deep - A boolean indicating whether the function is being called recursively for flow steps.
+	 * @param step_name - The name of the current step being processed in a flow.
+	 * @returns A promise that resolves to the updated context after all steps have been executed.
+	 * @throws {GlobalError} Throws a GlobalError if any step results in an error.
+	 */
+	async runSteps(ctx: Context, steps: NodeBase[], deep = false, step_name = ""): Promise<Context> {
 		ctx.config = { ...ctx.config };
 
 		try {
 			ctx.logger.log(`Starting runner for ${steps.length} steps ${!deep ? "(Parent)" : `(${step_name})`}`);
 			let flow = false;
-			let flow_steps: NanoService[] = [];
+			let flow_steps: NodeBase[] = [];
 			let flow_step = 0;
 			let stepName = "";
 
@@ -25,7 +34,7 @@ export default abstract class RunnerSteps {
 					if (ctx.response.error) throw ctx.response.error;
 				} else {
 					stepName = step.name;
-					flow_steps = (await step.processFlow(ctx)).data as NanoService[];
+					flow_steps = (await step.processFlow(ctx)).data as NodeBase[];
 
 					flow = true;
 					flow_step = i;